Set up govuk-frontend with Next.js
Followed the instructions at:
https://frontend.design-system.service.gov.uk/get-started/

Part of these instructions involve ensuring that the GOVUK assets
(images and fonts) are served from the appropriate static paths. Ideally
we'd handle this by adding the assets within node_modules as an extra
folder for Next.js to serve static files from - but after a brief bit of
research I can't see that Next (easily) supports the use of multiple
static/public folders.

Instead I've created a script to copy the files into the correct
location, and explicitly ignored those files from git. This script is
called as part of the npm `prepare` hook - so it will be called when
running an `npm install` (amongst other packaging operations).