Add missing newlines at end of files
We should probably investigate enforcing this with the linting in the
future. One suggestion was to use https://pre-commit.com/, but I'm
reluctant to introduce a Python dependency for a JS-only project. We can
instead probably write a custom hook for Husky that will do the same
thing.

For now, I've adjusted my editor settings to ensure that the trailing
newline character is always added to files.