Make local authentication provider return a JWT, not a User object
Following a discussion about how auth with Cognito will work in the
(near) future, it seems that it should be the responsibility of the
authentication provider to generate a token, not the general
user-service application.

Encapsulating the token generation inside the local authentication
provider follows the same functional pattern as getting Cognito to
generate a token for us.