Break out git hook installation into separate bash script
The `prepare` npm script is automatically run after `npm install`, even
if the `--production` flag is passed. As husky is a development
dependency, this resulted in an error being thrown after running `npm
install --production`.

Previously this was worked around by wrapping the call to husky in
a conditional that checked the node environment. Unfortunately this
doesn't play nicely with the CI, where it seems that the conditional
isn't being interpretted correctly.

This commit moves the call to husky, along with the checks for node
environment (and a new check for the CI environment), into a separate
script file where the interpeter is explicitly set to bash.