Ensure redirect to Bichard is using a 302, not a 307, HTTP redirect
Specifying `permanent: false` to the Next.js redirect object means that
under the hood Next.js sends a HTTP 307 redirect to the client. A 307
redirect specifies that the browser should use the same method to fetch
the new resource as it did for the previous one.

In our case, this is undesirable. The login form sends the credentials
to the user-service backend with a POST, but upon successful validation
of the credentials we want the user to get redirected to Bichard with
a GET request.

Fortunately, Next.js allows us to avoid the `permanent` syntax and
instead directly specify the status code to use.