Add basic Makefile for consistency with other bichard repos
For now, the targets in this Makefile are fairly basic - they simply
call the short examples given in the README, which are sufficient to
build/run the user-service.

This means that someone unfamiliar with the user-service, but with
experience in other Bichard repos, can come along and run a `make build`
or `make run` as they'd expect.