Pass --quiet-deps flag through to the Sass compiler
The GOV.UK Frontend Design System uses `/` for division in its SASS
stylesheets. Sass is deprecating the use of the `/` operator, in
preference of the `math.div()` function.

There is an open issue about this in the GOV.UK Frontend repo:

https://github.com/alphagov/govuk-frontend/issues/2238

In the meantime, passing the `--quiet-deps` to Sass in the user-service
app means we won't get flooded with warnings when running/building the
app.