Rip out all of the local authentication/users stuff
As part of the upcoming work for BICAWS-999, we're going to be removing
all "local" functionality of the user-service, in preference of just
requiring a database in all environments.

This will hopefully prevent the situation where we're adding lots of
logic and config to deal with edge-cases in the "local" code, when this
is intended to never be used in production anyway.