Rework email templating to support HTML and text versions
(Modern) emails typically have two main components to their body: a HTML
version and a text-only version. Previously we were only
rendering the text version as the body, but we're going to want to also
render and send HTML styled versions too.

This commit introduces some changes that allow emails to be rendered at
send-time using React and JSX.