Give each paragraph element a unique key
Next/React were complaining about the paragraph elements generated when
looking over the array with the following message:

  Warning: Each child in a list should have a unique "key" prop.

Adding a unique key to each element prevents this. In this case, the
only easily accessible unique property is the index of the string in the
array - which we've had to silence the eslint warning for.