Use constant number of iterations for password hashing
The number of iterations is used to determine the 'work' required to
calcualte the hash. This is common knowledge (it's plainly embedded at
the start of the password hash format), and we want to be certain about
enforcing the amount of work required to calculate a hash.

Therefore it doesn't make sense for this value to be randomised - we
don't gain anything from that besides having some hashes that are easier
to verify than others.