Remove --ignore-scripts from npm install in Docker build
This option was previously used because we didn't want to call our
`copy-govuk-frontend-assets.sh` and `install-git-hooks.sh` scripts when
installing the node dependencies in the final container build.

Now that we've moved away from the use of `prepare` in favour of
separate unique scripts that don't get automatically called, we no
longer need to disable the running of scripts when installing packages.