Ensure new users cannot be created with a CJSM email address
The logic in place in the application thus far implicitly expects the
user email address in the database to not end in .cjsm.net, and
removes/adds the .cjsm.net suffix to the email addresses at the relevant
entry/exit points from the system.

Weird things will almost certainly happen if a user is created with
a .cjsm.net email address.