Add flag for disabling the Secure cookie option
The Secure option, when set, means that the browser will only send the
cookies in the request when the request is being made over https.

Both `npm run dev` and `npm run start` result in Next.js serving the
content over http, not https - and so the login process does not work
because the cookies are not passed back from the browser to the server
after they are set.