Update asset references to use new asset path
The govuk-frontend npm package changed from the
`/assets/{fonts,images}/...` path to the `/{fonts,images}/...` path for
static assets in version 3.0.0.

We obviously missed this small print when updating the package at the
time, and thus the update broke our referencing of the assets - but for
anyone that didn't do an `npm install` after the version specified in
package.json was updated, it would have continued to work.